feat : 주제어, 지역코드 캐쉬 구현 - 주제어는 List, 지역코드는 Map 으로 구현

+++ src/main/java/com/steelrain/springboot/lilac/datamodel/LibraryRegionCodeDTO.java
@@ -18,7 +18,7 @@ public class LibraryRegionCodeDTO {
      */
     @NotNull
     @NotEmpty
-    private Integer code;
+    private Short code;
 
     @Size(max=2)
     @NotNull
